cephfs-pressure/requirements.txt
bcrypt==5.0.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.5
decorator==5.2.1
Deprecated==1.3.1
fabric==3.2.2
fqdn==1.5.1
invoke==2.2.1
paramiko==4.0.0
prometheus_client==0.24.1
pycparser==2.23
PyNaCl==1.6.2
PyYAML==6.0.3
typing_extensions==4.15.0
wrapt==2.1.2
